requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
httplib2==0.31.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.4
mypy_extensions==1.1.0
opentelemetry-api==1.37.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.4
pytest==8.4.2
python-dotenv==1.1.1
reportlab==4.4.4
requests==2.32.5
rsa==4.9.1
sentry-sdk==2.38.0
setuptools==80.9.0
tqdm==4.67.1
twilio==9.8.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.20.1
zipp==3.23.0
Flask-Migrate==4.0.7
python-dateutil 
